requirements.txt
attrs==21.2.0
autopep8==1.6.0
bracex==2.2.1
Cerberus==1.3.4
certifi==2021.10.8
charset-normalizer==2.0.9
ecs-logging==1.1.0
elastic-enterprise-search==7.15.0
elastic-transport==7.15.0
flake8==4.0.1
idna==3.3
importlib-metadata==4.2.0
iniconfig==1.1.1
ldap3==2.9.1
mccabe==0.6.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.6
pysmb==1.2.7
pytest==6.2.5
python-dateutil==2.8.2
python-ntlm3==1.0.2
PyYAML==6.0
requests==2.26.0
requests-ntlm==0.1.0
six==1.16.0
tika==1.24
toml==0.10.2
typing-extensions==4.0.1
urllib3==1.26.7
wcmatch==8.3
zipp==3.6.0
